+++ pysmartcocoon/__init__.py
@@ -3,5 +3,5 @@
 from .api import SmartCocoonAPI
 from .manager import SmartCocoonManager
 
-__version__ = "1.4.1"
+__version__ = "1.4.2"
 __all__ = ["SmartCocoonManager", "SmartCocoonAPI"]